requirements.txt
google-genai>=0.1.0
httpx[socks]
pysocks
watchdog
psutil
docker
chromadb